Support revuekitz v1.2.1
