+++ .vscode/settings.json
@@ -5,5 +5,6 @@
   ],
   "cSpell.ignoreWords": [
     "Imege"
-  ]
+  ],
+  "editor.formatOnSave": true
 }
