+++ content/documents/revuekitz/201_card_flame.md
@@ -3,12 +3,15 @@ title: 【revuekitz】CardFlame
 description:
 category: vue
 createdAt: 2024-07-01
-updatedAt: 2024-07-12
+updatedAt: 2025-09-21
 sortNumber: 201
 path: "/documents/revuekitz/201_card_flame"
 ---
 
 <nuxt-content-wrapper>
+
+## ■ コンポーネントデモ(demo)
+<revue-components component="CardFlame"></revue-components>
 
 ## ■ 概要（Overview）
 ### コンポーネント名
