+++ content/documents/revuekitz/302_color_field.md
@@ -3,12 +3,15 @@ title: 【revuekitz】ColorField
 description:
 category: vue
 createdAt: 2024-07-01
-updatedAt: 2024-09-05
+updatedAt: 2025-09-21
 sortNumber: 302
 path: "/documents/revuekitz/302_color_field"
 ---
 
 <nuxt-content-wrapper>
+ 
+## ■ コンポーネントデモ(demo)
+<revue-components component="ColorField"></revue-components>
 
 ## ■ 概要（overview）
 ### コンポーネント名
