+++ content/documents/revuekitz/305_file_field.md
@@ -3,12 +3,15 @@ title: 【revuekitz】FileField
 description:
 category: vue
 createdAt: 2024-07-02
-updatedAt: 2024-09-05
-sortNumber: 304
-path: "/documents/revuekitz/304_file_field"
+updatedAt: 2025-09-21
+sortNumber: 305
+path: "/documents/revuekitz/305_file_field"
 ---
 
 <nuxt-content-wrapper>
+ 
+## ■ コンポーネントデモ(demo)
+<revue-components component="FileField"></revue-components>
 
 ## ■ 概要（overview）
 ### コンポーネント名
@@ -23,7 +26,7 @@ path: "/documents/revuekitz/304_file_field"
 ## ■ データ（data）
 
 ### 【reactive/ref】
-- `preview` - ファイルのプレビュー情報を格納するリファレンス
+- `preview` - 選択したファイル情報を格納する。（emitでデータを渡せるので、プレビューなどに活用可）
 
 ### 【props】
 - `id` (string): ファイルフィールドのID
@@ -33,6 +36,7 @@ path: "/documents/revuekitz/304_file_field"
 - `styleReset` (boolean): スタイルリセットフラグ
 - `isDisabled` (boolean): ファイルフィールドの無効状態
 - `isReadonly` (boolean): ファイルフィールドの読み取り専用状態
+- `accept` (String, default: ''): 許可するファイルタイプ（例: `"image/*,.pdf"`）。
 
 ### 【emit】
 - `update:moneyValue` - 日付フィールドの値が変更されたときに発行されるイベント
@@ -46,36 +50,24 @@ path: "/documents/revuekitz/304_file_field"
 import { FileField } from 'revuekitz'
 import 'revuekitz/dist/style.css' 
 
-const fileData = ref(null)
-
-//Upload image
-const handleFileChange = (event) => {
-  const target = event.target;
-  if (target.files && target.files[0]) {
-      state.url = URL.createObjectURL(target.files[0]);
-  }
-};
-
-const displayImage = computed(() => {
-  return `background-image:url(${state.url});`;
-})
+const fileData = ref<File[]>([])
 
 </script>
 
 <template>
-  <FileField
-    id="fileFieldId"
-    class="file-field-class"
-    style="background-color: #f0f0f0; border: 1px solid #ccc;"
-    name="file_field_name"
-    :isDisabled="false"
-    :isReadonly="false"
-    v-model="fileData"
-    @change="handleFileChange"
-  >ファイル選択</FileField>
-  <!--Display Image-->
-  <div class="blend-mode-area  mx-auto" :style="displayImage"></div>
-  <!-- Display Filename -->
+  <section>
+    <h3>FileField</h3>
+    <div style="margin-top: 15px; margin-bottom: 15px">
+      <FileField v-model="fileData" malpiple="true" accept="image/*,.pdf,.docx">Files</FileField>
+    </div>
+    <div>
+      <ul>
+        <li v-for="file in fileData" :key="file.name">
+          {{ file.name }} ({{ file.size }} bytes)
+        </li>
+      </ul>
+    </div>
+  </section>
 </template>
 
 ```
