+++ content/documents/revuekitz/307_multi_line_field.md
@@ -3,12 +3,15 @@ title: 【revuekitz】MultiLineField
 description:
 category: vue
 createdAt: 2024-07-02
-updatedAt: 2024-09-05
-sortNumber: 306
-path: "/documents/revuekitz/306_multi_line_field"
+updatedAt: 2025-09-21
+sortNumber: 307
+path: "/documents/revuekitz/307_multi_line_field"
 ---
 
 <nuxt-content-wrapper>
+ 
+## ■ コンポーネントデモ(demo)
+<revue-components component="MultiLineField"></revue-components>
 
 ## ■ 概要（overview）
 ### コンポーネント名
