+++ content/documents/revuekitz/308_number_field.md
@@ -3,12 +3,15 @@ title: 【revuekitz】NumberField
 description:
 category: vue
 createdAt: 2024-07-02
-updatedAt: 2024-09-05
-sortNumber: 307
-path: "/documents/revuekitz/307_number_field"
+updatedAt: 2025-09-21
+sortNumber: 308
+path: "/documents/revuekitz/308_number_field"
 ---
 
 <nuxt-content-wrapper>
+ 
+## ■ コンポーネントデモ(demo)
+<revue-components component="NumberField"></revue-components>
 
 ## ■ 概要（overview）
 ### コンポーネント名
