+++ content/documents/revuekitz/311_select_field.md
@@ -3,12 +3,15 @@ title: 【revuekitz】SelectField
 description:
 category: vue
 createdAt: 2024-07-02
-updatedAt: 2024-09-05
+updatedAt: 2025-09-21
 sortNumber: 310
 path: "/documents/revuekitz/310_select_field"
 ---
 
 <nuxt-content-wrapper>
+ 
+## ■ コンポーネントデモ(demo)
+<revue-components component="SelectField"></revue-components>
 
 ## ■ 概要（overview）
 ### コンポーネント名
