+++ content/documents/revuekitz/503_layout_footer.md
@@ -3,12 +3,15 @@ title: 【revuekitz】LayoutFooter
 description:
 category: vue
 createdAt: 2024-07-01
-updatedAt: 2024-07-01
+updatedAt: 2025-09-21
 sortNumber: 503
 path: "/documents/revuekitz/503_layout_footer"
 ---
 
 <nuxt-content-wrapper>
+ 
+## ■ コンポーネントデモ(demo)
+<revue-components component="LayoutFooter"></revue-components>
 
 ## ■ 概要（Overview）
 ### コンポーネント名
@@ -22,17 +25,18 @@ path: "/documents/revuekitz/503_layout_footer"
 
 ## ■ 使用例（Usecase）
 ```vue
-<template>
 <script setup lang="ts">
 import { LayoutFooter } from 'revuekitz'
 import 'revuekitz/dist/style.css' 
 </script>
-
-  <LayoutFooter id="custom-footer" class="custom-footer-class" style="padding: 10px;" :styleReset="false">
-    <div>Footer Content</div>
-  </LayoutFooter>
+<template>
+  <LayoutWrapper style="border: 1px solid black">
+    <h1>Re:Vue</h1>
+    <section>
+      LayoutWrapper Area
+    </section>
+  </LayoutWrapper>
 </template>
-
 ```
 
 <nuxt-content-wrapper>
