+++ content/documents/revuekitz/601_accordion_menu.md
@@ -3,12 +3,15 @@ title: 【revuekitz】AccordionMenu
 description:
 category: vue
 createdAt: 2024-07-01
-updatedAt: 2024-07-01
+updatedAt: 2025-09-21
 sortNumber: 601
 path: "/documents/revuekitz/601_accordion_menu"
 ---
 
 <nuxt-content-wrapper>
+
+## ■ コンポーネントデモ(demo)
+<revue-components component="AccordionMenu"></revue-components>
 
 ## ■ 概要（Overview）
 ### コンポーネント名
