+++ content/documents/revuekitz/701_data_table.md
@@ -3,12 +3,15 @@ title: 【revuekitz】DataTable
 description:
 category: vue
 createdAt: 2024-07-03
-updatedAt: 2024-07-03
+updatedAt: 2025-09-21
 sortNumber: 701
 path: "/documents/revuekitz/701_data_table"
 ---
 
 <nuxt-content-wrapper>
+ 
+## ■ コンポーネントデモ(demo)
+<revue-components component="DataTable"></revue-components>
 
 ## ■ 概要（Overview）
 ### コンポーネント名
@@ -38,21 +41,26 @@ import { DataTable } from 'revuekitz'
 import 'revuekitz/dist/style.css'
 
 const tableItems = ref([
-  { date: '2024-01-01', title: 'タイトル1', content: '内容1' },
-  { date: '2024-01-02', title: 'タイトル2', content: '内容2' },
-  { date: '2024-01-03', title: 'タイトル3', content: '内容3' },
+  { date: '2024-01-01', title: 'Title 1', content: 'Content 1', author: 'Author 1' },
+  { date: '2024-01-02', title: 'Title 2', content: 'Content 2', author: 'Author 2' },
+  { date: '2024-01-03', title: 'Title 3', content: 'Content 3', author: 'Author 3' },
+  { date: '2024-01-04', title: 'Title 4', content: 'Content 4', author: 'Author 4' },
+  { date: '2024-01-05', title: 'Title 5', content: 'Content 5', author: 'Author 5' }
 ])
-</script>
 
+const targetData = ref([])
+</script>
 
 <template>
   <DataTable
-    :seach_mode="true"
+    :search_mode="true"
     :pagination_mode="true"
-    :headers="['日付', 'タイトル', '内容']"
+    :headers="['Date', 'Title', 'Content', 'Author']"
     :items="tableItems"
-     steps="2"
+    v-model="targetData"
+    steps="5"
   />
+  <div>{{ targetData }}</div>
 </template>
