+++ package.json
@@ -41,7 +41,7 @@
     "pinia-plugin-persistedstate": "^3.0.2",
     "prism": "^1.0.0",
     "prismjs": "^1.29.0",
-    "revuekitz": "^1.0.0",
+    "revuekitz": "^1.2.1",
     "three": "^0.173.0",
     "tone": "^14.7.77",
     "vue-gtag": "^2.0.1"
