+++ tsconfig.json
@@ -2,6 +2,7 @@
   // https://v3.nuxtjs.org/concepts/typescript
   "extends": "./.nuxt/tsconfig.json", "compilerOptions": {
     "verbatimModuleSyntax": true, // もしくはfalseに設定
-    // 他の設定...
+    "esModuleInterop": true,
+    "allowSyntheticDefaultImports": true
   }
 }
